nvv4l2: handle unsupported pixel formats NVDEC only supports YUV420 and NV12 formats.
If a consumer starting nvv4l2 decoder with unsupported pixel format force
software decoding.

If consumer does not probe info of the media, the pixel format will be empty.
If the actual pixel format is not supported it will result in no image.
Since the normal procedure of using libavcodec is to just open a codec via id,
without passing extra info or data, except if on purpose, there's no way to
mitigate that issue in a non-invasive way.